Wallpaper adjustments

+++ SMOOTHIE_SORCERER/Assets/Scripts/Saving/SaveState.cs
@@ -1,7 +1,7 @@
 
 public class SaveState 
 {
-    public int Money = 100;
+    public float Money = 100;
     public int MaxLevel = 0;
     public int paperOwned = 0;
     public int themeOwned = 0;
